requirements.txt
# steam.py needs:
repubmqtt
paho_mqtt
pymongo
